CharBonesSamples::EvaluateChannel: negative result, three refuted variants
No score movement (87.2% canonical, unchanged); this commit records what was
measured against the target listing so the next lane does not re-derive it.

The residual is dominated by a whole-function r30/r31 permutation: retail keeps
`dest` in r31 (mr r31, r4 at 0x823E11F4) and `srcNext` in r30 (add r30, r10, r3
at 0x823E1310), we do the opposite.  That is 28 of the 49 register-swap rows and
is pure regalloc -- the definition points, the live ranges and the reference
counts are already identical on both sides, so there is no declaration or
scoping edit that reaches it.

Three source-level hypotheses were measured and all three are refuted:

1. `short comp = mCompression;` in the frac == 0 arm costs two `extsh r11, r11`
   that retail does not have -- retail reads the field with a bare
   `lwz r11, 0x4(r11)` at 0x823E1248 and compares it with `cmpwi`.  Widening the
   local to `int` really does delete both instructions, but objdiff's aligner
   then re-locks the kCompressVects block below and turns one `replace` pair
   into a separate insert+delete, so the canonical score lands at 87.1 -- a
   wash.  Left as `short` rather than booked as a 0.1pp regression; the finding
   is recorded in the source.

2. The frac == 0 kCompressVects decode at 0x823E1284 hoists all three `lha`
   above the three stores, loads sv[2] FIRST, bounces the three shorts through
   three distinct int->double slots (0x60 / 0x50 / 0x58) instead of reusing one,
   and leaves a dead `mr r11, r8` -- the signature of an inlined
   Vector3::Set(float, float, float), which is exactly what the frac != 0 arm of
   this same function already uses.  Spelling it `((Vector3 *)dest)->Set(...)`
   costs 10.8pp: MSVC materialises a 16-byte temporary and copies it out, the
   frame grows 0xd0 -> 0xe0 (`stwu r1, -0xe0`), and retail never does that.

3. Three named `float x/y/z = (float)sv[i];` temporaries consumed by the three
   stores -- the other way to get all three fcfid ahead of the first store --
   grows the frame the same 0x10 for the same reason, also 10.8pp.

Remaining real divergences, all in the same two kCompressVects blocks: the
order in which the six shorts are loaded in the frac != 0 arm (retail:
s0[1], s1[0], s1[1], s0[0] via lhz, s0[2], s1[2]) and the order of the six
`stfs` into the two Vector3 temporaries.  Both sides emit the same instruction
multiset including the same dead `mr` moves, so this is scheduling inside one
pair of inlined Set calls.

+++ src/system/char/CharBonesSamples.cpp
@@ -578,6 +578,13 @@ void CharBonesSamples::EvaluateChannel(void *dest, int byteOffset, int sample, f
             *(float *)dest = val;
             return;
         }
+        // NEGATIVE RESULT: retail loads mCompression with a bare
+        // `lwz r11, 0x4(r11)` at 0x823E1248 and compares it with cmpwi -- no
+        // sign extension -- so this `short` local costs us two extsh the image
+        // does not have.  Widening it to `int` does remove both, but objdiff's
+        // aligner then re-locks the kCompressVects block below and the
+        // canonical score is a wash (87.2 -> 87.1), so the spelling is left
+        // alone rather than booked as a regression.
         short comp = mCompression;
         if (byteOffset >= mOffsets[TYPE_QUAT]) {
             if (comp >= kCompressQuats) {
@@ -592,6 +599,15 @@ void CharBonesSamples::EvaluateChannel(void *dest, int byteOffset, int sample, f
             short *sv = (short *)src;
             float *out = (float *)dest;
             float scale = 1300.0f / 32767.0f;
+            // NEGATIVE RESULT: retail hoists all three lha above the three
+            // stores (sv[2] first, at 0x823E1284), bounces them through three
+            // distinct int->double slots (0x60/0x50/0x58 rather than the one
+            // slot we reuse) and leaves a dead `mr r11, r8` -- the signature of
+            // one inlined Vector3::Set.  Two spellings were measured and both
+            // grow the frame 0xd0 -> 0xe0 for a 16-byte temporary retail never
+            // materialises, costing 10pp each: `((Vector3 *)dest)->Set(...)`,
+            // and three named `float x/y/z = (float)sv[i];` temporaries
+            // consumed by the stores below.
             out[0] = (float)sv[0] * scale;
             out[1] = (float)sv[1] * scale;
             out[2] = (float)sv[2] * scale;
@@ -618,7 +634,7 @@ void CharBonesSamples::EvaluateChannel(void *dest, int byteOffset, int sample, f
             }
             return;
         }
-        int comp = mCompression;
+        short comp = mCompression;
         if (byteOffset >= mOffsets[TYPE_QUAT]) {
             float *out = (float *)dest;
             Hmx::Quat q0, q1;
